# models/partner.go
package models

import "time"

type Partner struct {
	ID        string    `json:"id"`
	Name      string    `json:"name"`
	APIKey    string    `json:"api_key"`
	APISecret string    `json:"api_secret"`
	IsActive  bool      `json:"is_active"`
	CreatedAt time.Time `json:"created_at"`
}

type ActivatePartnerRequest struct {
	IsActive bool `json:"is_active"`
}